+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/UpstreamKitTest.groovy
@@ -9,7 +9,7 @@ import org.gradle.testkit.runner.TaskOutcome
  */
 class UpstreamKitTest extends AbstractKitTest {
 
-    public static final String GRADLE_VERSION = '6.8.1'
+    public static final String GRADLE_VERSION = '7.2'
 
     def "Check violation detection without cache task"() {
         setup:
@@ -34,7 +34,7 @@ class UpstreamKitTest extends AbstractKitTest {
 //        debug()
 
         when: "run task"
-        BuildResult result = runVer(GRADLE_VERSION, 'check')
+        BuildResult result = runVer(GRADLE_VERSION, 'check', '--warning-mode', 'all')
 
         then: "task successful"
         result.task(':check').outcome == TaskOutcome.SUCCESS